+++ lab_1/task1.py
@@ -1,13 +1,16 @@
 import os
 import logging
-import for_work_with_file 
-
 from enum import Enum
-from constants import PATH
+
+import for_work_with_file
+from constants1 import PATH
+
 
 """
 Module for working with file operations and data processing.
 """
+
+
 class VigenerMode(Enum):
     ENCRYPT = 1
     DECRYPT = 2
@@ -36,16 +39,19 @@ def vigener_cipher(text, key, mode: VigenerMode, alphabet):
                 char_index = alphabet.index(char)
                 if mode == VigenerMode.ENCRYPT:
                     cipher_char_index = (char_index + shift) % len(alphabet)
-                    correct_cipher_char_index = (cipher_char_index + 1) % len(alphabet)
+                    correct_cipher_char_index = (
+                        cipher_char_index + 1) % len(alphabet)
                 elif mode == VigenerMode.DECRYPT:
                     cipher_char_index = (char_index - shift) % len(alphabet)
-                    correct_cipher_char_index = (cipher_char_index - 1) % len(alphabet)
+                    correct_cipher_char_index = (
+                        cipher_char_index - 1) % len(alphabet)
                 result_text += alphabet[correct_cipher_char_index]
                 key_index = (key_index + 1) % key_length
             else:
                 result_text += char
     except Exception as e:
-        logging.error(f"An error occurred during Vigener cipher operation: {str(e)}")
+        logging.error(
+            f"An error occurred during Vigener cipher operation: {str(e)}")
     return result_text
 
 
@@ -63,21 +69,23 @@ def main():
     encrypted = json_data.get("encrypted", "")
     decrypted = json_data.get("decrypted", "")
     key = for_work_with_file.read_key_from_json_file(absolute_path + key_data)
-    alphabet_data = for_work_with_file.read_from_json_file(absolute_path + alphabet_1)
+    alphabet_data = for_work_with_file.read_from_json_file(
+        absolute_path + alphabet_1)
     if 'alphabet' in alphabet_data:
         alphabet = alphabet_data['alphabet']
     else:
         logging.error("Alphabet not found in alphabet.json")
         return
-    text = for_work_with_file.read_from_file( absolute_path + text_data)
+    text = for_work_with_file.read_from_file(absolute_path + text_data)
     encrypted_text = vigener_cipher(text, key, VigenerMode.ENCRYPT, alphabet)
     for_work_with_file.save_to_file(absolute_path + encrypted, encrypted_text)
 
-    encrypted_text = for_work_with_file.read_from_file(absolute_path + encrypted)
-    decrypted_text = vigener_cipher(encrypted_text, key, VigenerMode.DECRYPT, alphabet)
+    encrypted_text = for_work_with_file.read_from_file(
+        absolute_path + encrypted)
+    decrypted_text = vigener_cipher(
+        encrypted_text, key, VigenerMode.DECRYPT, alphabet)
     for_work_with_file.save_to_file(absolute_path + decrypted, decrypted_text)
 
 
 if __name__ == '__main__':
     main()
-
