+++ lab_1/task2/analis.json
@@ -132,4 +132,3 @@
         0.002325581395348837
     ]
 ]
-
