+++ lab_1/task2/paths.json
@@ -1,6 +1,6 @@
 {
-    "analis_data": "\\isb\\lab_1\\task2\\analis.json",
-    "encrypted": "\\isb\\lab_1\\task2\\encrypted_text.txt",
-    "decrypted": "\\isb\\lab_1\\task2\\decrypted_text.txt"
+    "analis_data": "\\lab_1\\task2\\analis.json",
+    "encrypted": "\\lab_1\\task2\\encrypted_text.txt",
+    "decrypted": "\\lab_1\\task2\\decrypted_text.txt"
   }
