+++ lab_1/task2/task2_key.py
@@ -1,35 +1,34 @@
 dictonary_letter_value = {
-"ы": " ",
-" ": "М",
-"н": "Х",
-"а": "4",
-"л": "У",
-"т": "r",
-"й": "7",
-"о": "1",
-"с": "Е",
-"г": "b",
-"р": "Д",
-"м": "Ф",
-"в": "a",
-"з": "Р",
-"к": "8",
-"ь": "А",
-"е": "О",
-"п": "2",
-"б": "5",
-"д": "c",
-"я": "Л",
-"ш": "Ы",
-"х": "Й",
-"ж": "П",
-"ю": "К",
-"щ": "Ь",
-"у": "t",
-"э": "Б",
-"ф": "И",
-"и": ">",
-"ч": "<",
-"ц": "Ч"
+    "ы": " ",
+    " ": "М",
+    "н": "Х",
+    "а": "4",
+    "л": "У",
+    "т": "r",
+    "й": "7",
+    "о": "1",
+    "с": "Е",
+    "г": "b",
+    "р": "Д",
+    "м": "Ф",
+    "в": "a",
+    "з": "Р",
+    "к": "8",
+    "ь": "А",
+    "е": "О",
+    "п": "2",
+    "б": "5",
+    "д": "c",
+    "я": "Л",
+    "ш": "Ы",
+    "х": "Й",
+    "ж": "П",
+    "ю": "К",
+    "щ": "Ь",
+    "у": "t",
+    "э": "Б",
+    "ф": "И",
+    "и": ">",
+    "ч": "<",
+    "ц": "Ч"
 }
-
